Update to 1.2.6
Resized the dialog to 480x500px. This should still fit on monitors that
are 1368x768. Therefore, the minimum resolution to run RPH Tools is now
that.
Tied the official blocking mechanism to RPH Tools. Blocking someone adds
them to the RPH Tools block list. Unblocking them will remove them.
Modified the chat room input textbox.
Added a placeholder that says "Posting as [username]". This will
disappear if you start typing something (and reappear if the box is
empty)
Changed the text color to the character's text color.